Update movement program and settings
The user can now move forward, left, right, and backwards using w,a,d,s respectfully. the right and left arrows are still used to pan right and left.  A maximum speed has been added to restrict the speed of the user. The maxSpeed and the acceleration can be altered in through Unity.